requirements.txt
asn1crypto==0.23.0
cachetools==2.0.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.3
enum34==1.1.6
fuzzywuzzy==0.15.1
google-api-python-client==1.6.4
google-auth==1.2.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.10.3
idna==2.6
ipaddress==1.0.18
oauth2client==4.1.2
oauthlib==2.0.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyOpenSSL==17.4.0
python-Levenshtein==0.12.0
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
